Avoid unnecessary copies
Pass some parameters by reference to avoid unnecessary copies.

When the parameter is intended to be copied into a member variable,
pass it by value and move it into the member instead.

This doesn't fix every occurrence of the issue but this diff is already
getting big so let's stop here for the moment.